I made the mixin work by myself! I was going to get help, but everyone was busy.  Does this mean that I'm Ruby potty-trained?  I sure hope so.
